Testsuite cleanups
